feat: add bigint support
This patch adds bigint support as the return value for the factorial
calculation.

ECMAScript is not very good with numbers, so we had to do add some
functionality to the bigint prototype to show the results.

The reasoning behind this was to make the experiment more interesting by
giving the server potentially more work to do - we can now trigger a
much larger calculation by asking the server to calculate the factorial
of n > 100 for example.